README.md
Practical Microsoft Server infrastructure project as a final exam assignment related to my education. The task involved setting up and documenting a complete Microsoft-based network environment for two merging companies. This included the installation and configuration of Active Directory, DHCP, DNS, DFS, Hyper-V, and Certificate Services, as well as the creation of appropriate access controls and group policies for various departments. The project also focused on security and backup solutions to meet the requirements of a modern business environment.